separate files for weekly schedule includes for #1891
