+++ IBPSA/Resources/C-Sources/WeeklySchedule.c
@@ -416,90 +416,4 @@ void* weeklyScheduleInit(const int tableOnFile, const char* name, const double t
   return (void*) scheduleID;
 }
 
-void weeklyScheduleFreeInit(void * ID) {
-  WeeklySchedule* scheduleID = (WeeklySchedule*)ID;
-
-  if (scheduleID->lastData != NULL)
-    free(scheduleID->lastData);
-
-  if (scheduleID->token != NULL)
-    free(scheduleID->token);
-
-  if (scheduleID->fp != NULL)
-    fclose(scheduleID->fp);
-
-  if (scheduleID->buff2 != NULL)
-    free(scheduleID->buff2);
-
-}
-
-
-void weeklyScheduleFree(void * ID) {
-  int i;
-  WeeklySchedule* scheduleID = (WeeklySchedule*)ID;
-
-  if (ID == NULL) /* Otherwise OM segfaults when IBPSA.Utilities.IO.Files.Examples.WeeklySchedule triggers an error */
-    return;
-
-  for (i = 0; i < scheduleID->n_allocatedRulesData; ++i) {
-    free(scheduleID->rules[i]->data);
-  }
- 
-  for (i = 0; i < scheduleID->n_allocatedRules; ++i) {
-    free(scheduleID->rules[i]);
-  }
-
-  if (scheduleID->rules != NULL)
-    free(scheduleID->rules);
-
-  free(ID);
-  ID = NULL;
-}
-
-/* Get a column value. Cache the last used row internally to speed up lookup. */
-double getScheduleValue(void * ID, const int column, const double modelicaTime) {
-  WeeklySchedule* scheduleID = (WeeklySchedule*)ID;
-  /* extrapolation for weeks that are outside of the user-defined range */
-  double t = modelicaTime - scheduleID->t_offset;
-  const double weekLen = 7 * 24 * 3600;
-  double time = fmod(t - weekLen * floor(t / weekLen), weekLen);
-  int i;
-  const int columnIndex = column - 1; /* Since we do not store the time indices in the data table */
-
-  /* Not calling weeklyScheduleFreeInit() or weeklyScheduleFree() since weeklyScheduleFreeInit() has already been called at the end of the
-  initialization and Modelica will call weeklyScheduleFree() upon a call of ModelicaFormatError) */
-  if (column < 0 || column > scheduleID->n_data_cols) {
-    ModelicaFormatError("The requested column index '%i' is outside of the table range '%i'.", column + 1, scheduleID->n_data_cols);
-  }
-  if (column == 0 ) {
-    ModelicaFormatError("The column index 1 is not a data column and is reserved for 'time'. It should not be read.");
-  }
-
-
-  if (time == scheduleID->previousTimestamp) {
-    i = scheduleID->previousIndex;
-  } else if (time > scheduleID->rules[scheduleID->previousIndex]->time) {
-    for (i = scheduleID->previousIndex; i < scheduleID->n_allocatedRules - 1; i ++) {
-      if (scheduleID->rules[i + 1]->time > time) {
-        break;
-      }
-    }
-  } else {
-    for (i = scheduleID->previousIndex; i > 0; i--) {
-      if (scheduleID->rules[i - 1]->time < time) {
-        i = i - 1;
-        break;
-      }
-    }
-    /* if time is smaller than the first row, wrap back to the end of the week */
-    if (i == 0 && scheduleID->rules[0]->time > time) {
-      i = scheduleID->n_allocatedRules - 1;
-    }
-  }
-  scheduleID->previousIndex = i;
-  scheduleID->previousTimestamp = time;
-
-  return scheduleID->rules[i]->data[columnIndex];
-}
-
 #endif
